requirements.txt
fastapi==0.104.1
boto3==1.29.7
langchain==0.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
